+++ mirp/settings/transformation_parameters.py
@@ -26,9 +26,9 @@ class ImageTransformationSettingsClass:
 
     Parameters
     ----------
-    by_slice: str or bool, optional, default: False
-        Defines whether calculations should be performed in 2D (True) or 3D (False), or alternatively only in the
-        largest slice ("largest"). See :class:`~mirp.settings.general_parameters.GeneralSettingsClass`.
+    by_slice: bool, optional, default: False
+        Defines whether calculations should be performed in 2D (True) or 3D (False).
+        See :class:`~mirp.settings.general_parameters.GeneralSettingsClass`.
 
     ibsi_compliant: bool, optional, default: True
         Limits use of filters to those that exist in the IBSI reference standard.